--- a/.cargo/config.toml
+++ b/.cargo/config.toml
@@ -1,2 +1,2 @@
 [build]
-rustflags = ["--emit=asm", "-C", "target-feature=+aes,+ssse3"]
+rustflags = ["--emit=asm"]