--- a/.github/workflows/rust-ci.yml
+++ b/.github/workflows/rust-ci.yml
@@ -1,9 +1,5 @@
 name: Rust Tests
 
-# Ask the Rust compiler to use AESNI
-env:
-  RUSTFLAGS: "-C target-feature=+aes,+ssse3"
-
 on:
   push:
     branches: