--- a/.github/workflows/rust-ci.yml
+++ b/.github/workflows/rust-ci.yml
@@ -95,7 +95,6 @@ jobs:
         with:
           name: Clippy
           token: ${{ secrets.GITHUB_TOKEN }}
-          # args: --all-features --all-targets -- -D warnings
           args: --all-features --all-targets
 
   codecov: