--- a/Cargo.toml
+++ b/Cargo.toml
@@ -10,7 +10,7 @@ readme = "README.md"
 
 [dependencies]
 maturin = "0.8.2"
-aesni = "0.10.0"
+aes = "0.7.4"
 rand = "0.7.3"
 rayon = "1.5.0"