## Description

Since the [`aesni`](https://docs.rs/aesni/0.10.0/aesni/) crate restricts the supported backends, this was causing some problems related to the usage of `SyMPC` on `aarch64` architectures (See https://github.com/OpenMined/SyMPC/issues/286). Replacing this crate with [`aes`](https://docs.rs/aes/0.7.4/aes/) will help because, by default, _"it performs runtime detection of CPU intrinsics and uses them if they are available"_.

Also, I removed the flags in the Cargo config:

* `--emit=asm` because it was causing problems with Clippy, while `cargo bench` can be run just fine without it.
* `-Ctarget-feature=+aes,+ssse3` because otherwise we would be forcing AES-NI as mentioned [here](https://docs.rs/aes/0.7.4/aes/#x86x86_64-intrinsics-aes-ni).


Closes #24 

## Affected Dependencies
aesni -> aes

## How has this been tested?
- cargo / pytest

## Checklist
- [x] I have followed the [Contribution Guidelines](https://github.com/OpenMined/.github/blob/master/CONTRIBUTING.md) and [Code of Conduct](https://github.com/OpenMined/.github/blob/master/CODE_OF_CONDUCT.md)
- [x] I have commented my code following the [OpenMined Styleguide](https://github.com/OpenMined/.github/blob/master/STYLEGUIDE.md)
- [x] I have labeled this PR with the relevant [Type labels](https://github.com/OpenMined/.github/labels?q=Type%3A)
- [x] My changes are covered by tests