@tholop 2021-07-31
Nice work Arturo! And great of we can simplify the crate and the compilation options at the same time. 

Please, do you have a benchmark that shows that the `aes` crate is indeed as fast as the `aesni` crate when the system supports AES-NI? Especially after preparing the package with Maturin. 

(I think that we have a pending branch/issue for some automated benchmarks with Criterion but I never finished it)