@arturomf94 2021-07-31
Currently we only have what's in `sycret/benches/`. The output is as follows:

With the changes:

```
Eq keygen               time:   [8.8554 us 8.8628 us 8.8739 us]                       
                        change: [-1.5175% -0.5167% +0.3938%] (p = 0.31 > 0.05)
                        No change in performance detected.
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) high mild
  6 (6.00%) high severe
```

Without the changes:

```
Eq keygen               time:   [6.1032 us 6.1108 us 6.1208 us]                       
                        change: [-31.143% -30.960% -30.808%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  2 (2.00%) high severe
```

So if I understand correctly, there is a slight decrease in performance? I do think it would be nice to have a more robust benchmark for comparison though. 

Btw... Ran these with: 

```
Architecture:        x86_64
CPU op-mode(s):      32-bit, 64-bit
Byte Order:          Little Endian
Address sizes:       39 bits physical, 48 bits virtual
CPU(s):              8
On-line CPU(s) list: 0-7
Thread(s) per core:  2
Core(s) per socket:  4
Socket(s):           1
NUMA node(s):        1
Vendor ID:           GenuineIntel
CPU family:          6
Model:               142
Model name:          Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
Stepping:            12
CPU MHz:             800.007
CPU max MHz:         4600.0000
CPU min MHz:         400.0000
BogoMIPS:            3984.00
Virtualization:      VT-x
L1d cache:           32K
L1i cache:           32K
L2 cache:            256K
L3 cache:            8192K
NUMA node0 CPU(s):   0-7
```