@tholop 2021-07-31
Indeed, thanks for the results. I'll try to add more benches to see if the performance drop is acceptable (it's a good time to do that, since we need benchmarks anyway). Let's keep the PR open for now!