@arturomf94 2021-09-15
@tholop what do you think? 

I've rebased this branch so that it runs the benchmarks. It seems to me that performance is only very slightly (insignificantly?) affected by the change. You can compare the results yourself: 

[master](https://pipelines.actions.githubusercontent.com/pWNCw8KPd5cszhteisbNzkPD7N2wc0dvhqIgj5J2Tb48vicVg0/_apis/pipelines/1/runs/215/signedlogcontent/14?urlExpires=2021-09-15T02%3A27%3A50.1097226Z&urlSigningMethod=HMACV1&urlSignature=rsr7ZpUFTEizS7bSwAvMkmj76v1HEZ1CaekKGbiyByg%3D) vs [PR](https://pipelines.actions.githubusercontent.com/pWNCw8KPd5cszhteisbNzkPD7N2wc0dvhqIgj5J2Tb48vicVg0/_apis/pipelines/1/runs/215/signedlogcontent/14?urlExpires=2021-09-15T02%3A27%3A50.1097226Z&urlSigningMethod=HMACV1&urlSignature=rsr7ZpUFTEizS7bSwAvMkmj76v1HEZ1CaekKGbiyByg%3D)