@tholop 2021-09-21
Thank you, Arturo! I double-checked the benchmarks on my laptop too, and indeed there is no significant performance hit.
I'm merging this, thanks for the PR and for the issues discovered and solved on the way :)